Fix all code
