+++ wasm-rpc/src/json.rs
@@ -12,13 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use bigdecimal::{BigDecimal, FromPrimitive, ToPrimitive};
-use golem_wasm_ast::analysis::{AnalysedFunctionParameter, AnalysedFunctionResult, AnalysedType};
+use golem_wasm_ast::analysis::{AnalysedFunctionParameter, AnalysedFunctionResult};
 use serde_json::{Number, Value as JsonValue};
-use std::collections::HashMap;
-use std::str::FromStr;
 
-use crate::{TypeAnnotatedValue, TypeAnnotatedValueResult, Uri, Value};
+use crate::{TypeAnnotatedValue, TypeAnnotatedValueResult, Value};
 
 pub fn function_parameters(
     value: &JsonValue,
@@ -33,8 +30,8 @@ pub fn function_parameters(
 
     if parameters.len() == expected_parameters.len() {
         for (json, fp) in parameters.iter().zip(expected_parameters.iter()) {
-            match validate_function_parameter(json, &fp.typ) {
-                Ok(result) => results.push(result),
+            match TypeAnnotatedValue::from_json_value(json, &fp.typ) {
+                Ok(result) => results.push(Value::from(result)),
                 Err(err) => errors.extend(err),
             }
         }
@@ -72,484 +69,6 @@ pub fn function_result(
             JsonValue::Object(map)
         }
     })
-}
-
-fn validate_function_parameter(
-    input_json: &JsonValue,
-    expected_type: &AnalysedType,
-) -> Result<Value, Vec<String>> {
-    match expected_type {
-        AnalysedType::Bool => get_bool(input_json),
-        AnalysedType::S8 => get_s8(input_json),
-        AnalysedType::U8 => get_u8(input_json),
-        AnalysedType::S16 => get_s16(input_json),
-        AnalysedType::U16 => get_u16(input_json),
-        AnalysedType::S32 => get_s32(input_json),
-        AnalysedType::U32 => get_u32(input_json),
-        AnalysedType::S64 => get_s64(input_json),
-        AnalysedType::U64 => get_u64(input_json).map(Value::U64),
-        AnalysedType::F64 => {
-            let num = bigdecimal(input_json)?;
-            let value = Value::F64(
-                num.to_string()
-                    .parse()
-                    .map_err(|err| vec![format!("Failed to parse f64: {}", err)])?,
-            );
-            Ok(value)
-        }
-        AnalysedType::F32 => get_f32(input_json),
-        AnalysedType::Chr => get_char(input_json).map(Value::Char),
-        AnalysedType::Str => get_string(input_json).map(Value::String),
-        AnalysedType::Enum(names) => get_enum(input_json, names).map(Value::Enum),
-        AnalysedType::Flags(names) => get_flag(input_json, names).map(Value::Flags),
-        AnalysedType::List(elem) => get_list(input_json, elem).map(Value::List),
-        AnalysedType::Option(elem) => get_option(input_json, elem).map(Value::Option),
-        AnalysedType::Result { ok, error } => get_result(input_json, ok, error).map(Value::Result),
-        AnalysedType::Record(fields) => get_record(input_json, fields).map(Value::Record),
-        AnalysedType::Variant(cases) => {
-            get_variant(input_json, cases).map(|result| Value::Variant {
-                case_idx: result.0,
-                case_value: result.1,
-            })
-        }
-        AnalysedType::Tuple(elems) => get_tuple(input_json, elems).map(Value::Tuple),
-        AnalysedType::Resource { .. } => get_handle(input_json),
-    }
-}
-
-fn get_bool(json: &JsonValue) -> Result<Value, Vec<String>> {
-    match json {
-        JsonValue::Bool(bool_val) => Ok(Value::Bool(*bool_val)),
-        _ => {
-            let type_description = type_description(json);
-            Err(vec![format!(
-                "Expected function parameter type is Boolean. But found {}",
-                type_description
-            )])
-        }
-    }
-}
-
-fn get_s8(json: &JsonValue) -> Result<Value, Vec<String>> {
-    ensure_range(
-        json,
-        BigDecimal::from_i8(i8::MIN).expect("Failed to convert i8::MIN to BigDecimal"),
-        BigDecimal::from_i8(i8::MAX).expect("Failed to convert i8::MAX to BigDecimal"),
-    )
-    .map(|num| Value::S8(num.to_i8().expect("Failed to convert BigDecimal to i8")))
-}
-
-fn get_u8(json: &JsonValue) -> Result<Value, Vec<String>> {
-    ensure_range(
-        json,
-        BigDecimal::from_u8(u8::MIN).expect("Failed to convert u8::MIN to BigDecimal"),
-        BigDecimal::from_u8(u8::MAX).expect("Failed to convert u8::MAX to BigDecimal"),
-    )
-    .map(|num| Value::U8(num.to_u8().expect("Failed to convert BigDecimal to u8")))
-}
-
-fn get_s16(json: &JsonValue) -> Result<Value, Vec<String>> {
-    ensure_range(
-        json,
-        BigDecimal::from_i16(i16::MIN).expect("Failed to convert i16::MIN to BigDecimal"),
-        BigDecimal::from_i16(i16::MAX).expect("Failed to convert i16::MAX to BigDecimal"),
-    )
-    .map(|num| Value::S16(num.to_i16().expect("Failed to convert BigDecimal to i16")))
-}
-
-fn get_u16(json: &JsonValue) -> Result<Value, Vec<String>> {
-    ensure_range(
-        json,
-        BigDecimal::from_u16(u16::MIN).expect("Failed to convert u16::MIN to BigDecimal"),
-        BigDecimal::from_u16(u16::MAX).expect("Failed to convert u16::MAX to BigDecimal"),
-    )
-    .map(|num| Value::U16(num.to_u16().expect("Failed to convert BigDecimal to u16")))
-}
-
-fn get_s32(json: &JsonValue) -> Result<Value, Vec<String>> {
-    ensure_range(
-        json,
-        BigDecimal::from_i32(i32::MIN).expect("Failed to convert i32::MIN to BigDecimal"),
-        BigDecimal::from_i32(i32::MAX).expect("Failed to convert i32::MAX to BigDecimal"),
-    )
-    .map(|num| Value::S32(num.to_i32().expect("Failed to convert BigDecimal to i32")))
-}
-
-fn get_u32(json: &JsonValue) -> Result<Value, Vec<String>> {
-    ensure_range(
-        json,
-        BigDecimal::from_u32(u32::MIN).expect("Failed to convert u32::MIN to BigDecimal"),
-        BigDecimal::from_u32(u32::MAX).expect("Failed to convert u32::MAX to BigDecimal"),
-    )
-    .map(|num| Value::U32(num.to_u32().expect("Failed to convert BigDecimal to u32")))
-}
-
-fn get_s64(json: &JsonValue) -> Result<Value, Vec<String>> {
-    ensure_range(
-        json,
-        BigDecimal::from_i64(i64::MIN).expect("Failed to convert i64::MIN to BigDecimal"),
-        BigDecimal::from_i64(i64::MAX).expect("Failed to convert i64::MAX to BigDecimal"),
-    )
-    .map(|num| Value::S64(num.to_i64().expect("Failed to convert BigDecimal to i64")))
-}
-
-fn get_f32(json: &JsonValue) -> Result<Value, Vec<String>> {
-    ensure_range(
-        json,
-        BigDecimal::from_f32(f32::MIN).expect("Failed to convert f32::MIN to BigDecimal"),
-        BigDecimal::from_f32(f32::MAX).expect("Failed to convert f32::MAX to BigDecimal"),
-    )
-    .map(|num| Value::F32(num.to_f32().expect("Failed to convert BigDecimal to f32")))
-}
-
-fn ensure_range(
-    value: &JsonValue,
-    min: BigDecimal,
-    max: BigDecimal,
-) -> Result<BigDecimal, Vec<String>> {
-    let num = bigdecimal(value)?;
-    if num >= min && num <= max {
-        Ok(num)
-    } else {
-        Err(vec![format!(
-            "value {} is not within the range of {} to {}",
-            value, min, max
-        )])
-    }
-}
-
-fn get_u64(value: &JsonValue) -> Result<u64, Vec<String>> {
-    match value {
-        JsonValue::Number(num) => {
-            if let Some(u64) = num.as_u64() {
-                Ok(u64)
-            } else {
-                Err(vec![format!("Cannot convert {} to u64", num)])
-            }
-        }
-        _ => {
-            let type_description = type_description(value);
-            Err(vec![format!(
-                "Expected function parameter type is u64. But found {}",
-                type_description
-            )])
-        }
-    }
-}
-
-fn bigdecimal(value: &JsonValue) -> Result<BigDecimal, Vec<String>> {
-    match value {
-        JsonValue::Number(num) => {
-            if let Ok(f64) = BigDecimal::from_str(num.to_string().as_str()) {
-                Ok(f64)
-            } else {
-                Err(vec![format!("Cannot convert {} to f64", num)])
-            }
-        }
-        _ => {
-            let type_description = type_description(value);
-            Err(vec![format!(
-                "Expected function parameter type is BigDecimal. But found {}",
-                type_description
-            )])
-        }
-    }
-}
-
-fn get_char(json: &JsonValue) -> Result<char, Vec<String>> {
-    if let Some(num_u64) = json.as_u64() {
-        if num_u64 > u32::MAX as u64 {
-            Err(vec![format!(
-                "The value {} is too large to be converted to a char",
-                num_u64
-            )])
-        } else {
-            char::from_u32(num_u64 as u32).ok_or(vec![format!(
-                "The value {} is not a valid unicode character",
-                num_u64
-            )])
-        }
-    } else {
-        let type_description = type_description(json);
-
-        Err(vec![format!(
-            "Expected function parameter type is Char. But found {}",
-            type_description
-        )])
-    }
-}
-
-fn get_string(input_json: &JsonValue) -> Result<String, Vec<String>> {
-    if let Some(str_value) = input_json.as_str() {
-        // If the JSON value is a string, return it
-        Ok(str_value.to_string())
-    } else {
-        // If the JSON value is not a string, return an error with type information
-        let type_description = type_description(input_json);
-        Err(vec![format!(
-            "Expected function parameter type is String. But found {}",
-            type_description
-        )])
-    }
-}
-
-fn type_description(value: &JsonValue) -> &'static str {
-    match value {
-        JsonValue::Null => "Null",
-        JsonValue::Bool(_) => "Boolean",
-        JsonValue::Number(_) => "Number",
-        JsonValue::String(_) => "String",
-        JsonValue::Array(_) => "Array",
-        JsonValue::Object(_) => "Object",
-    }
-}
-
-#[allow(clippy::type_complexity)]
-fn get_result(
-    input_json: &JsonValue,
-    ok_type: &Option<Box<AnalysedType>>,
-    err_type: &Option<Box<AnalysedType>>,
-) -> Result<Result<Option<Box<Value>>, Option<Box<Value>>>, Vec<String>> {
-    fn validate(
-        typ: &Option<Box<AnalysedType>>,
-        input_json: &JsonValue,
-    ) -> Result<Option<Box<Value>>, Vec<String>> {
-        if let Some(typ) = typ {
-            validate_function_parameter(input_json, typ).map(|v| Some(Box::new(v)))
-        } else if input_json.is_null() {
-            Ok(None)
-        } else {
-            Err(vec![
-                "The type of ok is absent, but some JSON value was provided".to_string(),
-            ])
-        }
-    }
-
-    match input_json.get("ok") {
-        Some(value) => Ok(Ok(validate(ok_type, value)?)),
-        None => match input_json.get("err") {
-            Some(value) => Ok(Err(validate(err_type, value)?)),
-            None => Err(vec![
-                "Failed to retrieve either ok value or err value".to_string()
-            ]),
-        },
-    }
-}
-
-fn get_option(
-    input_json: &JsonValue,
-    tpe: &AnalysedType,
-) -> Result<Option<Box<Value>>, Vec<String>> {
-    match input_json.as_null() {
-        Some(_) => Ok(None),
-
-        None => validate_function_parameter(input_json, tpe).map(|result| Some(Box::new(result))),
-    }
-}
-
-fn get_list(input_json: &JsonValue, tpe: &AnalysedType) -> Result<Vec<Value>, Vec<String>> {
-    let json_array = input_json
-        .as_array()
-        .ok_or(vec![format!("Input {} is not an array", input_json)])?;
-
-    let mut errors: Vec<String> = vec![];
-    let mut vals: Vec<Value> = vec![];
-
-    for json in json_array {
-        match validate_function_parameter(json, tpe) {
-            Ok(result) => vals.push(result),
-            Err(errs) => errors.extend(errs),
-        }
-    }
-
-    if errors.is_empty() {
-        Ok(vals)
-    } else {
-        Err(errors)
-    }
-}
-
-fn get_tuple(input_json: &JsonValue, types: &[AnalysedType]) -> Result<Vec<Value>, Vec<String>> {
-    let json_array = input_json.as_array().ok_or(vec![format!(
-        "Input {} is not an array representing tuple",
-        input_json
-    )])?;
-
-    if json_array.len() != types.len() {
-        return Err(vec![format!(
-            "The length of types in template is not equal to the length of tuple (array) in  {}",
-            input_json,
-        )]);
-    }
-
-    let mut errors: Vec<String> = vec![];
-    let mut vals: Vec<Value> = vec![];
-
-    for (json, tpe) in json_array.iter().zip(types.iter()) {
-        match validate_function_parameter(json, tpe) {
-            Ok(result) => vals.push(result),
-            Err(errs) => errors.extend(errs),
-        }
-    }
-
-    if errors.is_empty() {
-        Ok(vals)
-    } else {
-        Err(errors)
-    }
-}
-
-fn get_record(
-    input_json: &JsonValue,
-    name_type_pairs: &[(String, AnalysedType)],
-) -> Result<Vec<Value>, Vec<String>> {
-    let json_map = input_json.as_object().ok_or(vec![format!(
-        "The input {} is not a json object",
-        input_json
-    )])?;
-
-    let mut errors: Vec<String> = vec![];
-    let mut vals: Vec<Value> = vec![];
-
-    for (name, tpe) in name_type_pairs {
-        if let Some(json_value) = json_map.get(name) {
-            match validate_function_parameter(json_value, tpe) {
-                Ok(result) => vals.push(result),
-                Err(value_errors) => errors.extend(
-                    value_errors
-                        .iter()
-                        .map(|err| format!("Invalid value for the key {}. Error: {}", name, err))
-                        .collect::<Vec<_>>(),
-                ),
-            }
-        } else {
-            match tpe {
-                AnalysedType::Option(_) => vals.push(Value::Option(None)),
-                _ => errors.push(format!("Key '{}' not found in json_map", name)),
-            }
-        }
-    }
-
-    if errors.is_empty() {
-        Ok(vals)
-    } else {
-        Err(errors)
-    }
-}
-
-fn get_enum(input_json: &JsonValue, names: &[String]) -> Result<u32, Vec<String>> {
-    let input_enum_value = input_json
-        .as_str()
-        .ok_or(vec![format!("Input {} is not string", input_json)])?;
-
-    let mut discriminant: Option<i32> = None;
-
-    for (pos, name) in names.iter().enumerate() {
-        if input_enum_value == name {
-            discriminant = Some(pos as i32)
-        }
-    }
-
-    if let Some(d) = discriminant {
-        Ok(d as u32)
-    } else {
-        Err(vec![format!(
-            "Invalid input {}. Valid values are {}",
-            input_enum_value,
-            names.join(",")
-        )])
-    }
-}
-
-fn get_flag(input_json: &JsonValue, names: &[String]) -> Result<Vec<bool>, Vec<String>> {
-    let input_flag_values = input_json.as_array().ok_or(vec![format!(
-        "Input {} is not an array to be parsed as flags",
-        input_json
-    )])?;
-
-    let mut discriminant_map: HashMap<&str, usize> = HashMap::new();
-
-    for (pos, name) in names.iter().enumerate() {
-        discriminant_map.insert(name.as_str(), pos);
-    }
-
-    let mut result: Vec<bool> = vec![false; names.len()];
-
-    for i in input_flag_values {
-        let json_str = i
-            .as_str()
-            .ok_or(vec![format!("{} is not a valid string", i)])?;
-        if let Some(d) = discriminant_map.get(json_str) {
-            result[*d] = true;
-        } else {
-            return Err(vec![format!(
-                "Invalid input {}. It should be one of {}",
-                json_str,
-                names.join(",")
-            )]);
-        }
-    }
-
-    Ok(result)
-}
-
-fn get_variant(
-    input_json: &JsonValue,
-    types: &[(String, Option<AnalysedType>)],
-) -> Result<(u32, Option<Box<Value>>), Vec<String>> {
-    let mut possible_mapping_indexed: HashMap<&String, (usize, &Option<AnalysedType>)> =
-        HashMap::new();
-
-    for (pos, (name, optional_type)) in types.iter().enumerate() {
-        possible_mapping_indexed.insert(name, (pos, optional_type));
-    }
-
-    let json_obj = input_json
-        .as_object()
-        .ok_or(vec![format!("Input {} is not an object", input_json)])?;
-
-    let (key, json) = if json_obj.is_empty() {
-        Err(vec!["Zero variants in in the input".to_string()])
-    } else {
-        Ok(json_obj.iter().next().unwrap())
-    }?;
-
-    match possible_mapping_indexed.get(key) {
-        Some((index, Some(tpe))) => validate_function_parameter(json, tpe)
-            .map(|result| (*index as u32, Some(Box::new(result)))),
-        Some((index, None)) if json.is_null() => Ok((*index as u32, None)),
-        Some((_, None)) => Err(vec![format!("Unit variant {key} has non-null JSON value")]),
-        None => Err(vec![format!("Unknown key {key} in the variant")]),
-    }
-}
-
-fn get_handle(value: &JsonValue) -> Result<Value, Vec<String>> {
-    match value.as_str() {
-        Some(str) => {
-            // not assuming much about the url format, just checking it ends with a /<resource-id-u64>
-            let parts: Vec<&str> = str.split('/').collect();
-            if parts.len() >= 2 {
-                match u64::from_str(parts[parts.len() - 1]) {
-                    Ok(resource_id) => {
-                        let uri = parts[0..(parts.len() - 1)].join("/");
-                        Ok(Value::Handle { uri: Uri { value: uri }, resource_id })
-                    }
-                    Err(err) => {
-                        Err(vec![format!("Failed to parse resource-id section of the handle value: {}", err)])
-                    }
-                }
-            } else {
-                Err(vec![format!(
-                    "Expected function parameter type is Handle, represented by a worker-url/resource-id string. But found {}",
-                    str
-                )])
-            }
-        }
-        None => Err(vec![format!(
-            "Expected function parameter type is Handle, represented by a worker-url/resource-id string. But found {}",
-            type_description(value)
-        )]),
-    }
 }
 
 pub struct JsonFunctionResult(pub serde_json::Value);
@@ -684,7 +203,7 @@ impl From<TypeAnnotatedValue> for JsonFunctionResult {
 
 #[cfg(test)]
 mod tests {
-    use crate::json::{get_record, validate_function_parameter, JsonFunctionResult};
+    use crate::json::JsonFunctionResult;
     use crate::{TypeAnnotatedValue, Value};
     use golem_wasm_ast::analysis::AnalysedType;
     use proptest::prelude::*;
@@ -697,6 +216,13 @@ mod tests {
     ) -> Result<JsonValue, Vec<String>> {
         TypeAnnotatedValue::from_value(val, expected_type)
             .map(|result| JsonFunctionResult::from(result).0)
+    }
+
+    fn validate_function_parameter(
+        json: &JsonValue,
+        expected_type: &AnalysedType,
+    ) -> Result<Value, Vec<String>> {
+        TypeAnnotatedValue::from_json_value(json, expected_type).map(Value::from)
     }
 
     proptest! {
@@ -1153,11 +679,13 @@ mod tests {
             (key2.clone(), AnalysedType::Str),
         ];
 
-        let result = get_record(&input_json, &name_type_pairs);
+        let result =
+            TypeAnnotatedValue::get_record(&input_json, &name_type_pairs).map(|v| Value::from(v));
         let expected_result = Ok(vec![
             Value::String("value1".to_string()),
             Value::String("value2".to_string()),
         ]);
+
         assert_eq!(result, expected_result);
 
         // Test case where a key is missing
@@ -1165,7 +693,8 @@ mod tests {
             "key1": "value1",
         });
 
-        let result = get_record(&input_json, &name_type_pairs);
+        let result =
+            TypeAnnotatedValue::get_record(&input_json, &name_type_pairs).map(|v| Value::from(v));
         assert!(result.is_err());
     }
 }
