+++ wasm-rpc/src/lib.rs
@@ -48,15 +48,23 @@ pub mod wasmtime;
 use crate::builder::WitValueBuilder;
 pub use builder::{NodeBuilder, WitValueBuilderExtensions};
 pub use extractor::{WitNodePointer, WitValueExtractor};
+use std::collections::HashMap;
 use std::ops::Deref;
+use std::str::FromStr;
 
 #[cfg(feature = "host")]
 use ::wasmtime::component::bindgen;
 
 #[cfg(feature = "typeinfo")]
+use bigdecimal::{BigDecimal, FromPrimitive, ToPrimitive};
+
+#[cfg(feature = "typeinfo")]
 use golem_wasm_ast::analysis::{
     AnalysedFunctionResult, AnalysedResourceId, AnalysedResourceMode, AnalysedType,
 };
+
+#[cfg(feature = "typeinfo")]
+use serde_json::Value as JsonValue;
 
 #[cfg(feature = "host")]
 bindgen!({
@@ -707,6 +715,569 @@ impl TypeAnnotatedValue {
                 _ => Err(vec!["Unexpected type; expected a Handle type.".to_string()]),
             },
         }
+    }
+
+    pub fn from_json_value(
+        json_val: &JsonValue,
+        analysed_type: &AnalysedType,
+    ) -> Result<TypeAnnotatedValue, Vec<String>> {
+        match analysed_type {
+            AnalysedType::Bool => TypeAnnotatedValue::get_bool(json_val),
+            AnalysedType::S8 => TypeAnnotatedValue::get_s8(json_val),
+            AnalysedType::U8 => TypeAnnotatedValue::get_u8(json_val),
+            AnalysedType::S16 => TypeAnnotatedValue::get_s16(json_val),
+            AnalysedType::U16 => TypeAnnotatedValue::get_u16(json_val),
+            AnalysedType::S32 => TypeAnnotatedValue::get_s32(json_val),
+            AnalysedType::U32 => TypeAnnotatedValue::get_u32(json_val),
+            AnalysedType::S64 => TypeAnnotatedValue::get_s64(json_val),
+            AnalysedType::U64 => TypeAnnotatedValue::get_u64(json_val),
+            AnalysedType::F64 => TypeAnnotatedValue::get_f64(json_val),
+            AnalysedType::F32 => TypeAnnotatedValue::get_f32(json_val),
+            AnalysedType::Chr => TypeAnnotatedValue::get_char(json_val),
+            AnalysedType::Str => TypeAnnotatedValue::get_string(json_val),
+            AnalysedType::Enum(names) => TypeAnnotatedValue::get_enum(json_val, names),
+            AnalysedType::Flags(names) => TypeAnnotatedValue::get_flag(json_val, names),
+            AnalysedType::List(elem) => TypeAnnotatedValue::get_list(json_val, elem),
+            AnalysedType::Option(elem) => TypeAnnotatedValue::get_option(json_val, elem),
+            AnalysedType::Result { ok, error } => {
+                TypeAnnotatedValue::get_result(json_val, ok, error)
+            }
+            AnalysedType::Record(fields) => TypeAnnotatedValue::get_record(json_val, fields),
+            AnalysedType::Variant(cases) => TypeAnnotatedValue::get_variant(json_val, cases),
+            AnalysedType::Tuple(elems) => TypeAnnotatedValue::get_tuple(json_val, elems),
+            AnalysedType::Resource { id, resource_mode } => {
+                TypeAnnotatedValue::get_handle(json_val, id.clone(), resource_mode.clone())
+            }
+        }
+    }
+
+    fn get_bool(json: &JsonValue) -> Result<TypeAnnotatedValue, Vec<String>> {
+        match json {
+            JsonValue::Bool(bool_val) => Ok(TypeAnnotatedValue::Bool(*bool_val)),
+            _ => {
+                let type_description = type_description(json);
+                Err(vec![format!(
+                    "Expected function parameter type is Boolean. But found {}",
+                    type_description
+                )])
+            }
+        }
+    }
+
+    fn get_s8(json: &JsonValue) -> Result<TypeAnnotatedValue, Vec<String>> {
+        ensure_range(
+            json,
+            BigDecimal::from_i8(i8::MIN).expect("Failed to convert i8::MIN to BigDecimal"),
+            BigDecimal::from_i8(i8::MAX).expect("Failed to convert i8::MAX to BigDecimal"),
+        )
+        .map(|num| TypeAnnotatedValue::S8(num.to_i8().expect("Failed to convert BigDecimal to i8")))
+    }
+
+    fn get_u8(json: &JsonValue) -> Result<TypeAnnotatedValue, Vec<String>> {
+        ensure_range(
+            json,
+            BigDecimal::from_u8(u8::MIN).expect("Failed to convert u8::MIN to BigDecimal"),
+            BigDecimal::from_u8(u8::MAX).expect("Failed to convert u8::MAX to BigDecimal"),
+        )
+        .map(|num| TypeAnnotatedValue::U8(num.to_u8().expect("Failed to convert BigDecimal to u8")))
+    }
+
+    fn get_s16(json: &JsonValue) -> Result<TypeAnnotatedValue, Vec<String>> {
+        ensure_range(
+            json,
+            BigDecimal::from_i16(i16::MIN).expect("Failed to convert i16::MIN to BigDecimal"),
+            BigDecimal::from_i16(i16::MAX).expect("Failed to convert i16::MAX to BigDecimal"),
+        )
+        .map(|num| {
+            TypeAnnotatedValue::S16(num.to_i16().expect("Failed to convert BigDecimal to i16"))
+        })
+    }
+
+    fn get_u16(json: &JsonValue) -> Result<TypeAnnotatedValue, Vec<String>> {
+        ensure_range(
+            json,
+            BigDecimal::from_u16(u16::MIN).expect("Failed to convert u16::MIN to BigDecimal"),
+            BigDecimal::from_u16(u16::MAX).expect("Failed to convert u16::MAX to BigDecimal"),
+        )
+        .map(|num| {
+            TypeAnnotatedValue::U16(num.to_u16().expect("Failed to convert BigDecimal to u16"))
+        })
+    }
+
+    fn get_s32(json: &JsonValue) -> Result<TypeAnnotatedValue, Vec<String>> {
+        ensure_range(
+            json,
+            BigDecimal::from_i32(i32::MIN).expect("Failed to convert i32::MIN to BigDecimal"),
+            BigDecimal::from_i32(i32::MAX).expect("Failed to convert i32::MAX to BigDecimal"),
+        )
+        .map(|num| {
+            TypeAnnotatedValue::S32(num.to_i32().expect("Failed to convert BigDecimal to i32"))
+        })
+    }
+
+    fn get_u32(json: &JsonValue) -> Result<TypeAnnotatedValue, Vec<String>> {
+        ensure_range(
+            json,
+            BigDecimal::from_u32(u32::MIN).expect("Failed to convert u32::MIN to BigDecimal"),
+            BigDecimal::from_u32(u32::MAX).expect("Failed to convert u32::MAX to BigDecimal"),
+        )
+        .map(|num| {
+            TypeAnnotatedValue::U32(num.to_u32().expect("Failed to convert BigDecimal to u32"))
+        })
+    }
+
+    fn get_s64(json: &JsonValue) -> Result<TypeAnnotatedValue, Vec<String>> {
+        ensure_range(
+            json,
+            BigDecimal::from_i64(i64::MIN).expect("Failed to convert i64::MIN to BigDecimal"),
+            BigDecimal::from_i64(i64::MAX).expect("Failed to convert i64::MAX to BigDecimal"),
+        )
+        .map(|num| {
+            TypeAnnotatedValue::S64(num.to_i64().expect("Failed to convert BigDecimal to i64"))
+        })
+    }
+
+    fn get_f32(json: &JsonValue) -> Result<TypeAnnotatedValue, Vec<String>> {
+        ensure_range(
+            json,
+            BigDecimal::from_f32(f32::MIN).expect("Failed to convert f32::MIN to BigDecimal"),
+            BigDecimal::from_f32(f32::MAX).expect("Failed to convert f32::MAX to BigDecimal"),
+        )
+        .map(|num| {
+            TypeAnnotatedValue::F32(num.to_f32().expect("Failed to convert BigDecimal to f32"))
+        })
+    }
+
+    fn get_f64(json_val: &JsonValue) -> Result<TypeAnnotatedValue, Vec<String>> {
+        let num = bigdecimal(json_val)?;
+        let value = TypeAnnotatedValue::F64(
+            num.to_string()
+                .parse()
+                .map_err(|err| vec![format!("Failed to parse f64: {}", err)])?,
+        );
+        Ok(value)
+    }
+
+    fn get_string(json: &JsonValue) -> Result<TypeAnnotatedValue, Vec<String>> {
+        get_string(json).map(TypeAnnotatedValue::Str)
+    }
+
+    fn get_u64(json: &JsonValue) -> Result<TypeAnnotatedValue, Vec<String>> {
+        get_u64(json).map(TypeAnnotatedValue::U64)
+    }
+
+    fn get_char(json: &JsonValue) -> Result<TypeAnnotatedValue, Vec<String>> {
+        get_char(json).map(TypeAnnotatedValue::Chr)
+    }
+
+    fn get_tuple(
+        input_json: &JsonValue,
+        types: &[AnalysedType],
+    ) -> Result<TypeAnnotatedValue, Vec<String>> {
+        let json_array = input_json.as_array().ok_or(vec![format!(
+            "Input {} is not an array representing tuple",
+            input_json
+        )])?;
+
+        if json_array.len() != types.len() {
+            return Err(vec![format!(
+                "The length of types in template is not equal to the length of tuple (array) in  {}",
+                input_json,
+            )]);
+        }
+
+        let mut errors: Vec<String> = vec![];
+        let mut vals: Vec<TypeAnnotatedValue> = vec![];
+
+        for (json, tpe) in json_array.iter().zip(types.iter()) {
+            match TypeAnnotatedValue::from_json_value(json, tpe) {
+                Ok(result) => vals.push(result),
+                Err(errs) => errors.extend(errs),
+            }
+        }
+
+        if errors.is_empty() {
+            Ok(TypeAnnotatedValue::Tuple {
+                typ: types.to_vec(),
+                value: vals,
+            })
+        } else {
+            Err(errors)
+        }
+    }
+
+    fn get_option(
+        input_json: &JsonValue,
+        tpe: &AnalysedType,
+    ) -> Result<TypeAnnotatedValue, Vec<String>> {
+        match input_json.as_null() {
+            Some(_) => Ok(TypeAnnotatedValue::Option {
+                typ: tpe.clone(),
+                value: None,
+            }),
+
+            None => TypeAnnotatedValue::from_json_value(input_json, tpe).map(|result| {
+                TypeAnnotatedValue::Option {
+                    typ: tpe.clone(),
+                    value: Some(Box::new(result)),
+                }
+            }),
+        }
+    }
+
+    fn get_list(
+        input_json: &JsonValue,
+        tpe: &AnalysedType,
+    ) -> Result<TypeAnnotatedValue, Vec<String>> {
+        let json_array = input_json
+            .as_array()
+            .ok_or(vec![format!("Input {} is not an array", input_json)])?;
+
+        let mut errors: Vec<String> = vec![];
+        let mut vals: Vec<TypeAnnotatedValue> = vec![];
+
+        for json in json_array {
+            match TypeAnnotatedValue::from_json_value(json, tpe) {
+                Ok(result) => vals.push(result),
+                Err(errs) => errors.extend(errs),
+            }
+        }
+
+        if errors.is_empty() {
+            Ok(TypeAnnotatedValue::List {
+                typ: tpe.clone(),
+                values: vals,
+            })
+        } else {
+            Err(errors)
+        }
+    }
+
+    fn get_enum(
+        input_json: &JsonValue,
+        names: &[String],
+    ) -> Result<TypeAnnotatedValue, Vec<String>> {
+        let input_enum_value = input_json
+            .as_str()
+            .ok_or(vec![format!("Input {} is not string", input_json)])?;
+
+        if names.contains(&input_enum_value.to_string()) {
+            Ok(TypeAnnotatedValue::Enum {
+                typ: names.to_vec(),
+                value: input_enum_value.to_string(),
+            })
+        } else {
+            Err(vec![format!(
+                "Invalid input {}. Valid values are {}",
+                input_enum_value,
+                names.join(",")
+            )])
+        }
+    }
+
+    #[allow(clippy::type_complexity)]
+    fn get_result(
+        input_json: &JsonValue,
+        ok_type: &Option<Box<AnalysedType>>,
+        err_type: &Option<Box<AnalysedType>>,
+    ) -> Result<TypeAnnotatedValue, Vec<String>> {
+        fn validate(
+            typ: &Option<Box<AnalysedType>>,
+            input_json: &JsonValue,
+        ) -> Result<Option<Box<TypeAnnotatedValue>>, Vec<String>> {
+            if let Some(typ) = typ {
+                TypeAnnotatedValue::from_json_value(input_json, typ).map(|v| Some(Box::new(v)))
+            } else if input_json.is_null() {
+                Ok(None)
+            } else {
+                Err(vec![
+                    "The type of ok is absent, but some JSON value was provided".to_string(),
+                ])
+            }
+        }
+
+        match input_json.get("ok") {
+            Some(value) => Ok(TypeAnnotatedValue::Result {
+                ok: ok_type.clone(),
+                error: err_type.clone(),
+                value: Ok(validate(ok_type, value)?),
+            }),
+            None => match input_json.get("err") {
+                Some(value) => Ok(TypeAnnotatedValue::Result {
+                    ok: ok_type.clone(),
+                    error: err_type.clone(),
+                    value: Err(validate(err_type, value)?),
+                }),
+                None => Err(vec![
+                    "Failed to retrieve either ok value or err value".to_string()
+                ]),
+            },
+        }
+    }
+
+    fn get_record(
+        input_json: &JsonValue,
+        name_type_pairs: &[(String, AnalysedType)],
+    ) -> Result<TypeAnnotatedValue, Vec<String>> {
+        let json_map = input_json.as_object().ok_or(vec![format!(
+            "The input {} is not a json object",
+            input_json
+        )])?;
+
+        let mut errors: Vec<String> = vec![];
+        let mut vals: Vec<(String, TypeAnnotatedValue)> = vec![];
+
+        for (name, tpe) in name_type_pairs {
+            if let Some(json_value) = json_map.get(name) {
+                match TypeAnnotatedValue::from_json_value(json_value, tpe) {
+                    Ok(result) => vals.push((name.clone(), result)),
+                    Err(value_errors) => errors.extend(
+                        value_errors
+                            .iter()
+                            .map(|err| {
+                                format!("Invalid value for the key {}. Error: {}", name, err)
+                            })
+                            .collect::<Vec<_>>(),
+                    ),
+                }
+            } else {
+                match tpe {
+                    AnalysedType::Option(_) => vals.push((
+                        name.clone(),
+                        TypeAnnotatedValue::Option {
+                            typ: tpe.clone(),
+                            value: None,
+                        },
+                    )),
+                    _ => errors.push(format!("Key '{}' not found in json_map", name)),
+                }
+            }
+        }
+
+        if errors.is_empty() {
+            Ok(TypeAnnotatedValue::Record {
+                typ: name_type_pairs.to_vec(),
+                value: vals,
+            })
+        } else {
+            Err(errors)
+        }
+    }
+
+    fn get_flag(
+        input_json: &JsonValue,
+        names: &[String],
+    ) -> Result<TypeAnnotatedValue, Vec<String>> {
+        let json_array = input_json
+            .as_array()
+            .ok_or(vec![format!("Input {} is not an array", input_json)])?;
+
+        let mut errors: Vec<String> = vec![];
+        let mut vals: Vec<String> = vec![];
+
+        for json in json_array.iter() {
+            let flag: String = json
+                .as_str()
+                .map(|x| x.to_string())
+                .or_else(|| json.as_bool().map(|b| b.to_string()))
+                .or_else(|| json.as_number().map(|n| n.to_string()))
+                .ok_or(vec![format!(
+                    "Input {} is not a string or boolean or number",
+                    json
+                )])?;
+
+            if names.contains(&flag) {
+                vals.push(flag);
+            } else {
+                errors.push(format!(
+                    "Invalid input {}. Valid values are {}",
+                    flag,
+                    names.join(",")
+                ));
+            }
+        }
+
+        if errors.is_empty() {
+            Ok(TypeAnnotatedValue::Flags {
+                typ: names.to_vec(),
+                values: vals,
+            })
+        } else {
+            Err(errors)
+        }
+    }
+
+    fn get_variant(
+        input_json: &JsonValue,
+        types: &[(String, Option<AnalysedType>)],
+    ) -> Result<TypeAnnotatedValue, Vec<String>> {
+        let mut possible_mapping_indexed: HashMap<&String, &Option<AnalysedType>> = HashMap::new();
+
+        for (name, optional_type) in types.iter() {
+            possible_mapping_indexed.insert(name, optional_type);
+        }
+
+        let json_obj = input_json
+            .as_object()
+            .ok_or(vec![format!("Input {} is not an object", input_json)])?;
+
+        let (key, json) = if json_obj.is_empty() {
+            Err(vec!["Zero variants in in the input".to_string()])
+        } else {
+            Ok(json_obj.iter().next().unwrap())
+        }?;
+
+        match possible_mapping_indexed.get(key) {
+            Some(Some(tpe)) => TypeAnnotatedValue::from_json_value(json, tpe).map(|result| {
+                TypeAnnotatedValue::Variant {
+                    typ: types.to_vec(),
+                    case_name: key.clone(),
+                    case_value: Some(Box::new(result)),
+                }
+            }),
+            Some(None) if json.is_null() => Ok(TypeAnnotatedValue::Variant {
+                typ: types.to_vec(),
+                case_name: key.clone(),
+                case_value: None,
+            }),
+            Some(None) => Err(vec![format!("Unit variant {key} has non-null JSON value")]),
+            None => Err(vec![format!("Unknown key {key} in the variant")]),
+        }
+    }
+
+    fn get_handle(
+        value: &JsonValue,
+        id: AnalysedResourceId,
+        resource_mode: AnalysedResourceMode,
+    ) -> Result<TypeAnnotatedValue, Vec<String>> {
+        match value.as_str() {
+            Some(str) => {
+                // not assuming much about the url format, just checking it ends with a /<resource-id-u64>
+                let parts: Vec<&str> = str.split('/').collect();
+                if parts.len() >= 2 {
+                    match u64::from_str(parts[parts.len() - 1]) {
+                        Ok(resource_id) => {
+                            let uri = parts[0..(parts.len() - 1)].join("/");
+                            Ok(TypeAnnotatedValue::Handle { id, resource_mode, uri: Uri { value: uri }, resource_id })
+                        }
+                        Err(err) => {
+                            Err(vec![format!("Failed to parse resource-id section of the handle value: {}", err)])
+                        }
+                    }
+                } else {
+                    Err(vec![format!(
+                        "Expected function parameter type is Handle, represented by a worker-url/resource-id string. But found {}",
+                        str
+                    )])
+                }
+            }
+            None => Err(vec![format!(
+                "Expected function parameter type is Handle, represented by a worker-url/resource-id string. But found {}",
+                type_description(value)
+            )]),
+        }
+    }
+}
+
+fn type_description(value: &JsonValue) -> &'static str {
+    match value {
+        JsonValue::Null => "Null",
+        JsonValue::Bool(_) => "Boolean",
+        JsonValue::Number(_) => "Number",
+        JsonValue::String(_) => "String",
+        JsonValue::Array(_) => "Array",
+        JsonValue::Object(_) => "Object",
+    }
+}
+
+fn ensure_range(
+    value: &JsonValue,
+    min: BigDecimal,
+    max: BigDecimal,
+) -> Result<BigDecimal, Vec<String>> {
+    let num = bigdecimal(value)?;
+    if num >= min && num <= max {
+        Ok(num)
+    } else {
+        Err(vec![format!(
+            "value {} is not within the range of {} to {}",
+            value, min, max
+        )])
+    }
+}
+
+fn bigdecimal(value: &JsonValue) -> Result<BigDecimal, Vec<String>> {
+    match value {
+        JsonValue::Number(num) => {
+            if let Ok(f64) = BigDecimal::from_str(num.to_string().as_str()) {
+                Ok(f64)
+            } else {
+                Err(vec![format!("Cannot convert {} to f64", num)])
+            }
+        }
+        _ => {
+            let type_description = type_description(value);
+            Err(vec![format!(
+                "Expected function parameter type is BigDecimal. But found {}",
+                type_description
+            )])
+        }
+    }
+}
+
+fn get_u64(value: &JsonValue) -> Result<u64, Vec<String>> {
+    match value {
+        JsonValue::Number(num) => {
+            if let Some(u64) = num.as_u64() {
+                Ok(u64)
+            } else {
+                Err(vec![format!("Cannot convert {} to u64", num)])
+            }
+        }
+        _ => {
+            let type_description = type_description(value);
+            Err(vec![format!(
+                "Expected function parameter type is u64. But found {}",
+                type_description
+            )])
+        }
+    }
+}
+
+fn get_char(json: &JsonValue) -> Result<char, Vec<String>> {
+    if let Some(num_u64) = json.as_u64() {
+        if num_u64 > u32::MAX as u64 {
+            Err(vec![format!(
+                "The value {} is too large to be converted to a char",
+                num_u64
+            )])
+        } else {
+            char::from_u32(num_u64 as u32).ok_or(vec![format!(
+                "The value {} is not a valid unicode character",
+                num_u64
+            )])
+        }
+    } else {
+        let type_description = type_description(json);
+
+        Err(vec![format!(
+            "Expected function parameter type is Char. But found {}",
+            type_description
+        )])
+    }
+}
+
+fn get_string(input_json: &JsonValue) -> Result<String, Vec<String>> {
+    if let Some(str_value) = input_json.as_str() {
+        // If the JSON value is a string, return it
+        Ok(str_value.to_string())
+    } else {
+        // If the JSON value is not a string, return an error with type information
+        let type_description = type_description(input_json);
+        Err(vec![format!(
+            "Expected function parameter type is String. But found {}",
+            type_description
+        )])
     }
 }
